FIX sqapi doc
git-svn-id: svn://tron.homeunix.org/simutrans/simutrans/trunk@10844 8aca7d54-2c30-db11-9de9-000461428c89

+++ src/simutrans/script/api/api_schedule.cc
@@ -243,7 +243,7 @@ void export_schedule(HSQUIRRELVM vm)
 	 * In order to change this value for a convoy
 	 * call @ref convoy_x::change_schedule.
 	 */
-	uint8 current;
+	integer current;
 #else
 	create_slot(vm, "entries", 0);
 	create_slot(vm, "waytype", 0);
